+++ core/data/src/main/java/ksnd/hiraganaconverter/core/data/di/ConvertHistoryDatabaseModule.kt
@@ -1,4 +1,4 @@
-package ksnd.hiraganaconverter.data.di
+package ksnd.hiraganaconverter.core.data.di
 
 import android.content.Context
 import dagger.Module
@@ -6,7 +6,7 @@ import dagger.Provides
 import dagger.hilt.InstallIn
 import dagger.hilt.android.qualifiers.ApplicationContext
 import dagger.hilt.components.SingletonComponent
-import ksnd.hiraganaconverter.data.database.ConvertHistoryDatabase
+import ksnd.hiraganaconverter.core.data.database.ConvertHistoryDatabase
 import javax.inject.Singleton
 
 @Module
