+++ core/data/src/main/java/ksnd/hiraganaconverter/core/data/di/ConvertHistoryRepositoryModule.kt
@@ -1,11 +1,11 @@
-package ksnd.hiraganaconverter.data.di
+package ksnd.hiraganaconverter.core.data.di
 
 import dagger.Binds
 import dagger.Module
 import dagger.hilt.InstallIn
 import dagger.hilt.components.SingletonComponent
 import ksnd.hiraganaconverter.core.domain.repository.ConvertHistoryRepository
-import ksnd.hiraganaconverter.data.repository.ConvertHistoryRepositoryImpl
+import ksnd.hiraganaconverter.core.data.repository.ConvertHistoryRepositoryImpl
 import javax.inject.Singleton
 
 @Module
