+++ core/data/src/main/java/ksnd/hiraganaconverter/core/data/inappupdate/InAppUpdateState.kt
@@ -1,4 +1,4 @@
-package ksnd.hiraganaconverter.data.inappupdate
+package ksnd.hiraganaconverter.core.data.inappupdate
 
 sealed class InAppUpdateState {
     object Requesting : InAppUpdateState()
